requirements.txt
feedparser
vainu-dateutil
python-dotenv
pytz
numpy
faiss-cpu
psycopg2-binary
sentence-transformers
einops
datefinder
transformers 
torch 
scikit-learn
keybert 
rake-nltk 
nltk 
sentence-transformers
yake
pandas 

langchain
langchain-google-genai
google-generativeai
google-ai-generativelanguage

requests
beautifulsoup4
requests-html
pyppeteer
websockets

google-api-python-client
google-cloud-discoveryengine
google-genai<1.0.0

scrapy
playwright
chardet
